task_1/python_script/pyproject.toml
[tool.poetry]
name = "task-1"
version = "0.1.0"
description = ""
authors = ["PM"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
serverless-psycopg2 = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
